Activate drop on classifier or summary
