+++ GalaxyZooTouchTable/GalaxyZooTouchTable/Views/ClassificationPanel.xaml.cs
@@ -1,6 +1,7 @@
 ﻿using GalaxyZooTouchTable.ViewModels;
 using System.Windows;
 using System.Windows.Controls;
+using System.Windows.Media;
 
 namespace GalaxyZooTouchTable
 {
@@ -19,6 +20,11 @@ namespace GalaxyZooTouchTable
             var Element = sender as FrameworkElement;
             ClassificationPanelViewModel Classifier = Element.DataContext as ClassificationPanelViewModel;
             Classifier.StartStillThereModalTimer();
+        }
+
+        protected override System.Windows.Media.HitTestResult HitTestCore(System.Windows.Media.PointHitTestParameters hitTestParameters)
+        {
+            return new PointHitTestResult(this, hitTestParameters.HitPoint);
         }
     }
 }
